requirements.txt
flask==3.0.1
Flask-Cors==4.0.0
marshmallow==3.20.2
PyJWT==2.8.0